+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- add `panic_section` method to `HookBuilder` for overriding the printer for
+  the panic message at the start of panic reports
 
 ## [0.5.2] - 2020-08-31
 ### Added